LocalDispatcher : Terminate entire process tree on Windows
- When run via `subprocess`, the Gaffer wrapper is a child process of a Windows `cmd.exe` process. In that case, `os.killpg` does not terminate all child processes. Using Windows' `TASKKILL /T` seems to be the only reliable option without install adding modules like `psutil`.

+++ python/GafferDispatch/LocalDispatcher.py
@@ -295,7 +295,10 @@ class LocalDispatcher( GafferDispatch.Dispatcher ) :
 			while process.poll() is None :
 
 				if batch.blindData().get( "killed" ) :
-					os.killpg( process.pid, signal.SIGTERM )
+					if os.name == "nt" :
+						subprocess.Popen( "TASKKILL /F /PID {} /T".format( process.pid ), stdout = sys.stderr )
+					else :
+						os.killpg( process.pid, signal.SIGTERM )
 					self.__reportKilled( batch )
 					return False
